requirements.txt
asgiref==3.9.1
boto3==1.40.19
botocore==1.40.19
Django==5.2.5
django-storages==1.14.6
djangorestframework==3.16.1
jmespath==1.0.1
pillow==11.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
